refactor files and code
