+++ src/Frontend/test/preconditions/index.ts
@@ -5,9 +5,7 @@ export { hasServiceControlMonitoringInstanceUrl } from "../preconditions/hasServ
 export { hasUpToDateServiceControl } from "../preconditions/hasUpToDateServiceControl";
 export { hasUpToDateServicePulse } from "../preconditions/hasUpToDateServicePulse";
 export { errorsDefaultHandler } from "../preconditions/hasNoErrors";
-export { hasCustomChecksEmpty } from "./hasCustomChecksEmpty";
-export { hasCustomChecksFailing } from "./hasCustomChecksFailing";
-export { hasCustomChecksPassing } from "./hasCustomChecksPassing";
+export { hasCustomChecksEmpty, hasCustomChecksFailing, hasCustomChecksPassing } from "./customChecks";
 export { hasNoDisconnectedEndpoints } from "../preconditions/hasNoDisconnectedEndpoints";
 export { hasNoMonitoredEndpoints, hasMonitoredEndpointsList, monitoredEndpointsNamed } from "../preconditions/hasMonitoredEndpoints";
 export { hasEventLogItems } from "../preconditions/hasEventLogItems";
